@@ -94,63 +103,68 @@
   },
   "homepage": "https://joelpurra.com/projects/talkie/",
   "devDependencies": {
-    "@google-cloud/translate": "^0.8.1",
-    "babel-eslint": "^7.2.3",
+    "@google-cloud/translate": "^1.1.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.2",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-preset-env": "^1.5.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
-    "bluebird": "^3.5.0",
+    "bluebird": "^3.5.1",
     "chokidar-cli": "^1.2.0",
     "chrome-webstore-manager": "^0.4.1",
-    "concurrently": "^3.4.0",
+    "concurrently": "^3.5.1",
     "configvention": "^1.1.0",
     "ent": "^2.2.0",
-    "eslint": "^3.12.2",
+    "eslint": "^4.11.0",
     "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-compat": "^1.0.2",
-    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-compat": "^2.1.0",
+    "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-no-unsafe-innerhtml": "^1.0.16",
-    "eslint-plugin-node": "^5.0.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.5.1",
     "eslint-plugin-standard": "^3.0.1",
     "json": "^9.0.4",
-    "jsonfile": "^3.0.0",
+    "jsonfile": "^4.0.0",
+    "merge-options": "^1.0.0",
     "pre-commit": "^1.2.2",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-redux": "^5.0.5",
-    "redux": "^3.6.0",
-    "redux-persist": "=4.8.0",
-    "redux-persist-crosstab": "=3.5.3",
+    "prop-types": "^15.6.0",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
-    "rimraf": "^2.5.4",
-    "rollup": "^0.42.0",
-    "rollup-plugin-babel": "^2.7.1",
-    "rollup-plugin-buble": "^0.15.0",
-    "rollup-plugin-cleanup": "^1.0.0",
-    "rollup-plugin-commonjs": "^8.0.2",
+    "reselect": "^3.0.1",
+    "rimraf": "^2.6.2",
+    "rollup": "^0.51.8",
+    "rollup-plugin-babel": "^3.0.2",
+    "rollup-plugin-buble": "^0.18.0",
+    "rollup-plugin-commonjs": "^8.2.6",
+    "rollup-plugin-ejs": "^1.1.1",
+    "rollup-plugin-filesize": "^1.5.0",
     "rollup-plugin-json": "^2.1.0",
     "rollup-plugin-jsx": "^1.0.3",
-    "rollup-plugin-license": "^0.4.0",
+    "rollup-plugin-license": "^0.5.0",
     "rollup-plugin-node-globals": "^1.1.0",
     "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-replace": "^1.1.1",
-    "rollup-watch": "^4.0.0",
+    "rollup-plugin-replace": "^2.0.0",
+    "rollup-plugin-uglify": "^2.0.1",
+    "rollup-watch": "^4.3.1",
     "rrssb": "^1.14.0",
     "stream-to-promise": "^2.2.0",
-    "striptags": "^3.0.1",
-    "web-ext": "^1.8.1",
-    "webextension-polyfill": "^0.1.1"
+    "striptags": "^3.1.0",
+    "styletron-client": "^3.0.0-rc.3",
+    "styletron-react": "^3.0.0-rc.3",
+    "uglify-es": "^3.1.10",
+    "web-ext": "^2.2.2",
+    "webextension-polyfill": "^0.2.1"
   },
   "browserslist": [
     "last 5 Chrome versions",
     "last 5 Firefox versions"
   ],
-  "dependencies": {
-    "prop-types": "^15.5.10"
-  }
+  "dependencies": {}
 }
