@@ -21,81 +21,107 @@ along with Talkie.  If not, see <https://www.gnu.org/licenses/>.
 import React from "react";
 import PropTypes from "prop-types";
 
-import {
-    scrollIntoViewIfNeeded,
-} from "../../../utils/select-element";
+import styled from "../../../../shared/hocs/styled.jsx";
 
+import * as formBase from "../../../../shared/styled/form/form-base.jsx";
+
+import MultilineSelect from "../../../../shared/components/form/multiline-select.jsx";
+
+import translateAttribute from "../../../../shared/hocs/translate.jsx";
+
+@translateAttribute
 export default class AvailableVoices extends React.PureComponent {
     constructor(props) {
         super(props);
 
-        this.selectElement = null;
-
         this.handleChange = this.handleChange.bind(this);
+
+        this.styled = {
+            effectiveVoiceOption: styled({
+                fontWeight: "bold",
+            })(formBase.option),
+        };
     }
 
     static defaultProps = {
         voices: [],
         value: null,
-        defaultVoiceNameForSelectedLanguage: null,
+        effectiveVoiceNameForSelectedLanguage: null,
         disabled: true,
     };
 
     static propTypes = {
         voices: PropTypes.arrayOf(PropTypes.shape({
-            name: PropTypes.string.isRequired,
+            default: PropTypes.bool.isRequired,
             lang: PropTypes.string.isRequired,
+            localService: PropTypes.bool.isRequired,
+            name: PropTypes.string.isRequired,
+            voiceURI: PropTypes.string.isRequired,
         })).isRequired,
         value: PropTypes.string,
-        defaultVoiceNameForSelectedLanguage: PropTypes.string,
+        effectiveVoiceNameForSelectedLanguage: PropTypes.string,
         onChange: PropTypes.func.isRequired,
-        disabled: PropTypes.bool,
+        disabled: PropTypes.bool.isRequired,
+        translate: PropTypes.func.isRequired,
     };
 
-    handleChange(e) {
-        const voiceName = e.target.value;
-
+    handleChange(voiceName) {
         this.props.onChange(voiceName);
-
-        scrollIntoViewIfNeeded(this.selectElement);
     }
 
     render() {
+        const {
+            translate,
+        } = this.props;
+
+        const translatedVoiceFeatureOnline = translate("frontend_voiceFeatureOnline");
+
         const voicesOptions = this.props.voices.map(
             (voice) => {
-                const isDefaultVoiceNameForSelectedLanguage = (
-                    this.props.defaultVoiceNameForSelectedLanguage
-                    && this.props.defaultVoiceNameForSelectedLanguage === voice.name
+                const isEffectiveVoiceNameForSelectedLanguage = (
+                    this.props.effectiveVoiceNameForSelectedLanguage
+                    && this.props.effectiveVoiceNameForSelectedLanguage === voice.name
                 );
 
-                const className = isDefaultVoiceNameForSelectedLanguage ? "effective-voice" : "";
+                const VoiceOption = isEffectiveVoiceNameForSelectedLanguage ? this.styled.effectiveVoiceOption : formBase.option;
 
-                return <option
+                let voiceNameAndFeaturesText = voice.name;
+
+                const voiceFeatures = [];
+
+                if (voice.localService === false) {
+                    voiceFeatures.push(translatedVoiceFeatureOnline);
+                }
+
+                if (voiceFeatures.length > 0) {
+                    voiceNameAndFeaturesText += " (";
+                    voiceNameAndFeaturesText += voiceFeatures.join(", ");
+                    voiceNameAndFeaturesText += ")";
+                }
+
+                if (isEffectiveVoiceNameForSelectedLanguage) {
+                    voiceNameAndFeaturesText += " ✓";
+                }
+
+                return <VoiceOption
                     key={voice.name}
                     // TODO: proper way to store/look up objects?
                     value={voice.name}
-                    className={className}
-                       >
-                    {voice.name}{isDefaultVoiceNameForSelectedLanguage ? " ✓" : ""}
-                </option>;
+                >
+                    {voiceNameAndFeaturesText}
+                </VoiceOption>;
             }
         );
 
         return (
-            <select
-                id="voices-voices-list"
-                size="7"
+            <MultilineSelect
+                size={7}
                 onChange={this.handleChange}
                 value={this.props.value}
                 disabled={this.props.disabled}
-                ref={
-                    (selectElement) => {
-                        this.selectElement = selectElement;
-                        scrollIntoViewIfNeeded(this.selectElement);
-                    }}
             >
                 {voicesOptions}
-            </select>
+            </MultilineSelect>
         );
     }
 }
