@@ -31,20 +31,37 @@ import {
 } from "../shared/tabs";
 
 import {
-    knownEvents,
-} from "../shared/events";
-
-import {
     eventToPromise,
-    startFrontend,
-    stopFrontend,
+    startReactFrontend,
+    stopReactFrontend,
 } from "../frontend/shared-frontend";
+
+import loadRoot from "./load-root.jsx";
 
 import DualLogger from "../frontend/dual-log";
 
 const dualLogger = new DualLogger("popup.js");
 
-const shouldPassClick = () => {
+const passClickToBackground = () => promiseTry(
+    () => {
+        dualLogger.dualLogDebug("Start", "passClickToBackground");
+
+        return getBackgroundPage()
+            .then((background) => background.iconClick())
+            .then(() => {
+                dualLogger.dualLogDebug("Done", "passClickToBackground");
+
+                return undefined;
+            })
+            .catch((error) => {
+                dualLogger.dualLogError("passClickToBackground", error);
+
+                throw error;
+            });
+    }
+);
+
+const shouldPassClickOnLoad = () => {
     const containsBlockerString = document.location
         && typeof document.location.search === "string"
         && document.location.search.length > 0
@@ -53,111 +70,34 @@ const shouldPassClick = () => {
     return !containsBlockerString;
 };
 
-const passClickToBackground = () => promiseTry(
-        () => {
-            dualLogger.dualLogDebug("Start", "passClickToBackground");
+const passClickToBackgroundOnLoad = () => promiseTry(
+    () => {
+        // NOTE: provide a way to link to the popup without triggering the "click".
+        if (!shouldPassClickOnLoad()) {
+            dualLogger.dualLogDebug("Skipped", "passClickToBackgroundOnLoad");
 
-            // NOTE: provide a way to link to the popup without triggering the "click".
-            if (!shouldPassClick()) {
-                dualLogger.dualLogDebug("Skipped", "passClickToBackground");
-
-                return;
-            }
-
-            return getBackgroundPage()
-                .then((background) => background.iconClick())
-                .then(() => {
-                    dualLogger.dualLogDebug("Done", "passClickToBackground");
-
-                    return undefined;
-                })
-                .catch((error) => {
-                    dualLogger.dualLogError("passClickToBackground", error);
-
-                    throw error;
-                });
+            return;
         }
-);
 
-const registerStartStopButtonClickHandlers = () => promiseTry(
-    () => {
-        const startStopButtons = Array.from(document.querySelectorAll(":scope .start-stop-button"));
-
-        startStopButtons.forEach((optionsLink) => {
-            optionsLink.onclick = (event) => {
-                event.preventDefault();
-
-                getBackgroundPage()
-                        .then((background) => background.iconClick())
-                        .catch((error) => {
-                            dualLogger.dualLogError("registerStartStopButtonClickHandlers", "onclick", error);
-                        });
-
-                return false;
-            };
-        });
-
-        return undefined;
-    }
-);
-
-const updateProgress = (data) => promiseTry(
-    () => {
-        const progressBar = document.getElementById("progress");
-        progressBar.max = data.max - data.min;
-        progressBar.value = data.current;
-
-        return undefined;
-    }
-);
-
-const killSwitches = [];
-
-const executeKillSwitches = () => {
-    // NOTE: expected to have only synchronous methods for the relevant parts.
-    killSwitches.forEach((killSwitch) => {
-        try {
-            killSwitch();
-        } catch (error) {
-            dualLogger.dualLogError("executeKillSwitches", error);
-        }
-    });
-};
-
-const registerBroadcastListeners = () => promiseTry(
-    () => {
-        return getBackgroundPage()
-            .then((background) => background.broadcaster().registerListeningAction(knownEvents.updateProgress, (/* eslint-disable no-unused-vars*/actionName/* eslint-enable no-unused-vars*/, actionData) => updateProgress(actionData)))
-            .then((killSwitch) => killSwitches.push(killSwitch));
+        return passClickToBackground();
     }
 );
 
 const start = () => promiseTry(
-    () => {
-        return Promise.resolve()
-            .then(() => startFrontend(killSwitches))
-            .then(() => registerBroadcastListeners())
-            .then(() => registerStartStopButtonClickHandlers())
-            .then(() => passClickToBackground());
-    }
+    () => Promise.resolve()
+        .then(() => startReactFrontend())
+        .then(() => passClickToBackgroundOnLoad())
+        .then(() => loadRoot())
+        .then(() => undefined)
 );
 
 const stop = () => promiseTry(
-    () => {
-        // NOTE: probably won't be correctly executed as unload doesn't allow asynchronous calls.
-        return Promise.resolve()
-            .then(() => stopFrontend());
-    }
+    // NOTE: probably won't be correctly executed as before/unload doesn't guarantee asynchronous calls.
+    () => stopReactFrontend()
+        .then(() => undefined)
 );
 
 registerUnhandledRejectionHandler();
 
 document.addEventListener("DOMContentLoaded", eventToPromise.bind(null, start));
-
-// NOTE: probably won't be correctly executed as unload doesn't allow asynchronous calls.
-window.addEventListener("unload", eventToPromise.bind(null, stop));
-
-// NOTE: synchronous version.
-window.addEventListener("unload", () => {
-    executeKillSwitches();
-});
+window.addEventListener("beforeunload", eventToPromise.bind(null, stop));
