@@ -20,11 +20,11 @@ along with Talkie.  If not, see <https://www.gnu.org/licenses/>.
 
 import {
     promiseTry,
-} from "../shared/promise";
+} from "./promise";
 
 import {
     getBackgroundPage,
-} from "../shared/tabs";
+} from "./tabs";
 
 export const openUrlInNewTab = (url) => promiseTry(
     () => {
@@ -37,7 +37,29 @@ export const openUrlInNewTab = (url) => promiseTry(
             throw new Error("Bad url, only https:// allowed: " + url);
         }
 
-        browser.tabs.create({active: true, url: url});
+        return browser.tabs.create({
+            active: true,
+            url: url,
+        });
+    }
+);
+
+export const openInternalUrlInNewTab = (url) => promiseTry(
+    () => {
+        if (typeof url !== "string") {
+            throw new Error("Bad url: " + url);
+        }
+
+        // NOTE: only root-relative internal urls.
+        // NOTE: double-slash is protocol relative, checking just in case.
+        if (!url.startsWith("/") || url[1] === "/") {
+            throw new Error("Bad url, only internally rooted allowed: " + url);
+        }
+
+        return browser.tabs.create({
+            active: true,
+            url: url,
+        });
     }
 );
 
@@ -53,10 +75,31 @@ export const openUrlFromConfigurationInNewTab = (id) => promiseTry(
         })
 );
 
+export const openInternalUrlFromConfigurationInNewTab = (id) => promiseTry(
+    () => getBackgroundPage()
+        .then((background) => background.getConfigurationValue(`urls.${id}`))
+        .then((url) => {
+            if (typeof url !== "string") {
+                throw new Error("Bad url for id: " + id);
+            }
+
+            return openInternalUrlInNewTab(url);
+        })
+);
+
 export const openShortKeysConfiguration = () => promiseTry(
     () => {
         const url = "chrome://extensions/configureCommands";
 
-        return browser.tabs.create({active: true, url: url});
+        return browser.tabs.create({
+            active: true,
+            url: url,
+        });
+    }
+);
+
+export const openOptionsPage = () => promiseTry(
+    () => {
+        return browser.runtime.openOptionsPage();
     }
 );
